backend/src/requirements.txt
awsgi==0.0.5
Flask==3.1.2
flask_cors==6.0.1
openai==1.107.1
python-dotenv==1.1.1
